+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changes for riot-parser
 
 ### v0.0.2
+- Minor code reduction.
+- Added suport for SVG en the tests.
+- Added test/builder/tree-builder2.js as sample.
 - Matching literal regexes is a bit faster now.
 - Fixes incorrect regex that matches literal regexes.